+++ src/scenes/MenuScene.h
@@ -12,6 +12,6 @@ public:
   ~MenuScene();
 
   void processEvent(const SDL_Event &event) override;
-  void update(Uint32 delta) override;
+  void update(float delta) override;
   void render(SDL_Renderer *renderer) override;
 };
