+++ src/scenes/Scene.h
@@ -11,6 +11,6 @@ public:
 
   // returns true if event was processed. 
   virtual void processEvent(const SDL_Event &event) = 0;
-  virtual void update(Uint32 delta) = 0;
+  virtual void update(float delta) = 0;
   virtual void render(SDL_Renderer *renderer) = 0;
 };
