+++ src/utils/Timer.cpp
@@ -1,12 +1,12 @@
 #include "Timer.h"
 
-void Timer::update(Uint32 delta) {
+void Timer::update(float delta) {
   if (elapsed < timeout) {
     elapsed += delta;
   }
 }
 
-void Timer::reset(Uint32 newTimeout) {
+void Timer::reset(float newTimeout) {
   elapsed = 0;
   if (newTimeout) {
     timeout = newTimeout;
