+++ src/utils/Timer.h
@@ -1,16 +1,14 @@
 #pragma once
 
-#include <SDL.h>
-
 class Timer {
 private:
-  Uint32 elapsed = 0, timeout = 0;
+  float elapsed = 0, timeout = 0;
 
 public:
   bool hasTimedOut() const {
     return timeout <= elapsed;
   }
 
-  void update(Uint32 delta);
-  void reset(Uint32 newTimeout = 0);
+  void update(float delta);
+  void reset(float newTimeout = 0);
 };
